requirements.txt
pandas==1.2.3
SQLAlchemy==1.3.23
google-cloud-bigquery==2.16.0